Report basically done
It still needs spellchecking and perhaps a bit more stuff added in intro&conclusions and if there is time we should add a little sidebar with links to the various sections as well as to the github. Also it would be nice to have some sort of different background or something to make it more visually pleasing.
But as far as i'm concerned I've done all I could content-wise